+++ src/dipdup/config/abi_etherscan.py
@@ -1,5 +1,7 @@
 from typing import Literal
 
+from pydantic import ConfigDict
+from pydantic import Extra
 from pydantic.dataclasses import dataclass
 
 from dipdup.config import AbiDatasourceConfig
@@ -8,7 +10,7 @@ from dipdup.config import HttpConfig
 DEFAULT_ETHERSCAN_URL = 'https://api.etherscan.io/api'
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class EtherscanDatasourceConfig(AbiDatasourceConfig):
     """Etherscan datasource config