+++ src/dipdup/config/evm_node.py
@@ -2,15 +2,16 @@ from __future__ import annotations
 
 from typing import Literal
 
+from pydantic import ConfigDict
+from pydantic import Extra
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 
 from dipdup.config import HttpConfig
 from dipdup.config import IndexDatasourceConfig
-from dipdup.exceptions import ConfigurationError
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class EvmNodeDatasourceConfig(IndexDatasourceConfig):
     """Subsquid datasource config
 
@@ -34,11 +35,11 @@ class EvmNodeDatasourceConfig(IndexDatasourceConfig):
     @validator('url')
     def _valid_url(cls, v: str) -> str:
         if not v.startswith(('http://', 'https://')):
-            raise ConfigurationError('Ethereum node URL must start with http(s)://')
+            raise ValueError('Ethereum node URL must start with http(s)://')
         return v
 
     @validator('ws_url')
     def _valid_ws_url(cls, v: str | None) -> str | None:
         if v and not v.startswith(('ws://', 'wss://')):
-            raise ConfigurationError('Ethereum node WebSocket URL must start with ws(s)://')
+            raise ValueError('Ethereum node WebSocket URL must start with ws(s)://')
         return v