+++ src/dipdup/config/tezos_tzkt_events.py
@@ -4,6 +4,8 @@ from dataclasses import field
 from typing import TYPE_CHECKING
 from typing import Literal
 
+from pydantic import ConfigDict
+from pydantic import Extra
 from pydantic.dataclasses import dataclass
 
 from dipdup.config import HandlerConfig
@@ -20,7 +22,7 @@ if TYPE_CHECKING:
     from dipdup.subscriptions import Subscription
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class TzktEventsHandlerConfig(HandlerConfig):
     """Event handler config
 
@@ -48,7 +50,7 @@ class TzktEventsHandlerConfig(HandlerConfig):
         yield 'event', f'TzktEvent[{event_cls}]'
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class TzktEventsUnknownEventHandlerConfig(HandlerConfig):
     """Unknown event handler config
 
@@ -71,7 +73,7 @@ class TzktEventsUnknownEventHandlerConfig(HandlerConfig):
 TzktEventsHandlerConfigU = TzktEventsHandlerConfig | TzktEventsUnknownEventHandlerConfig
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class TzktEventsIndexConfig(TzktIndexConfig):
     """Event index config