+++ src/dipdup/config/tezos_tzkt_operations.py
@@ -6,6 +6,8 @@ from typing import Any
 from typing import Literal
 from typing import cast
 
+from pydantic import ConfigDict
+from pydantic import Extra
 from pydantic.dataclasses import dataclass
 
 from dipdup.config import CodegenMixin
@@ -28,7 +30,7 @@ if TYPE_CHECKING:
     from dipdup.subscriptions import Subscription
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class SubgroupIndexMixin:
     """`subgroup_index` field to track index of operation in group
 
@@ -127,7 +129,7 @@ class TezosPatternConfig(CodegenMixin):
         return arg_name, 'TzktOperationData'
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class OperationsHandlerTransactionPatternConfig(TezosPatternConfig, SubgroupIndexMixin):
     """Transaction handler pattern config
 
@@ -187,7 +189,7 @@ class OperationsHandlerTransactionPatternConfig(TezosPatternConfig, SubgroupInde
         return None
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class OperationsHandlerOriginationPatternConfig(TezosPatternConfig, SubgroupIndexMixin):
     """Origination handler pattern config
 
@@ -236,7 +238,7 @@ class OperationsHandlerOriginationPatternConfig(TezosPatternConfig, SubgroupInde
         return None
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class OperationsHandlerSmartRollupExecutePatternConfig(TezosPatternConfig, SubgroupIndexMixin):
     """Operation handler pattern config
 
@@ -273,7 +275,7 @@ class OperationsHandlerSmartRollupExecutePatternConfig(TezosPatternConfig, Subgr
         return None
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class TzktOperationsIndexConfig(TzktIndexConfig):
     """Operation index config
 
@@ -337,7 +339,7 @@ OperationsHandlerPatternConfigU = (
 )
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class TzktOperationsHandlerConfig(HandlerConfig):
     """Operation handler config
 
@@ -369,7 +371,7 @@ class TzktOperationsHandlerConfig(HandlerConfig):
             yield arg, arg_type
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class OperationUnfilteredHandlerConfig(HandlerConfig):
     """Handler of unfiltered operation index
 
@@ -386,7 +388,7 @@ class OperationUnfilteredHandlerConfig(HandlerConfig):
         yield 'operation', 'TzktOperationData'
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class TzktOperationsUnfilteredIndexConfig(TzktIndexConfig):
     """Operation index config