+++ src/dipdup/config/tezos_tzkt_token_transfers.py
@@ -3,6 +3,8 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from typing import Literal
 
+from pydantic import ConfigDict
+from pydantic import Extra
 from pydantic.dataclasses import dataclass
 from pydantic.fields import Field
 
@@ -19,7 +21,7 @@ if TYPE_CHECKING:
     from dipdup.subscriptions import Subscription
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class TzktTokenTransfersHandlerConfig(HandlerConfig):
     """Token transfer handler config
 
@@ -32,8 +34,7 @@ class TzktTokenTransfersHandlerConfig(HandlerConfig):
 
     contract: TezosContractConfig | None = None
     token_id: int | None = None
-    # FIXME: Can't use `from_` field alias in dataclass
-    # FIXME: See https://github.com/pydantic/pydantic/issues/4286 (fixed in upcoming v2)
+    # FIXME: Can't use `from_` field alias in dataclasses (fixed in `next` with Pydantic v2)
     from_: TezosContractConfig | None = None
     to: TezosContractConfig | None = None
 
@@ -47,7 +48,7 @@ class TzktTokenTransfersHandlerConfig(HandlerConfig):
         yield 'token_transfer', 'TzktTokenTransferData'
 
 
-@dataclass
+@dataclass(config=ConfigDict(extra=Extra.forbid), kw_only=True)
 class TzktTokenTransfersIndexConfig(TzktIndexConfig):
     """Token transfer index config