+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+## Fixed
+
+- config: Forbid extra arguments in config mappings.
+- config: Fixed issue with `from` filter being ignored.
+- config: Don't raise `ConfigurationError` from some model validators.
+
 ## [7.5.3] - 2024-03-28
 
 ### Fixed